Fixing e2e tests
