+++ testing/cli-e2etest/tracetestcli/exec.go
@@ -2,33 +2,51 @@ package tracetestcli
 
 import (
 	"fmt"
+	"os"
 	"strings"
 
 	"github.com/kubeshop/tracetest/cli-e2etest/command"
 )
 
 const (
-	tracetestCommand = "tracetest"
+	defaultTracetestCommand = "tracetest"
 )
 
-type CLIExecCommand func(string, ...ExecOption) (*command.ExecResult, error)
+type ExecOption func(*executionState)
+
+type executionState struct {
+	cliConfigFile string
+}
 
 func Exec(tracetestSubCommand string, options ...ExecOption) (*command.ExecResult, error) {
-	state := composeExecutionState(options...)
+	state := &executionState{}
+	for _, option := range options {
+		option(state)
+	}
 
 	if state.cliConfigFile != "" {
 		// append config at the start of the command
 		tracetestSubCommand = fmt.Sprintf("-c %s %s", state.cliConfigFile, tracetestSubCommand)
 	}
 
+	tracetestCommand := getTracetestCommand()
 	tracetestSubCommands := strings.Split(tracetestSubCommand, " ")
 
 	return command.Exec(tracetestCommand, tracetestSubCommands...)
 }
 
-func GetComposedExecFunc(cliCommand string) CLIExecCommand {
-	return func(subCommand string, opts ...ExecOption) (*command.ExecResult, error) {
-		fullSubCommand := fmt.Sprintf("%s %s", cliCommand, subCommand)
-		return Exec(fullSubCommand, opts...)
+func getTracetestCommand() string {
+	tracetestCommand := os.Getenv("TRACETEST_COMMAND")
+
+	if tracetestCommand == "" {
+		return defaultTracetestCommand
+	}
+
+	return tracetestCommand
+}
+
+func WithCLIConfig(cliConfig string) ExecOption {
+	return func(es *executionState) {
+		es.cliConfigFile = cliConfig
 	}
 }
