--- testing/cli-e2etest/tracetestcli/get/main.go
@@ -1,7 +1,0 @@
-package get
-
-import (
-	"github.com/kubeshop/tracetest/cli-e2etest/tracetestcli"
-)
-
-var Exec = tracetestcli.GetComposedExecFunc("get")
