--- testing/cli-e2etest/tracetestcli/help_test.go
@@ -1,24 +1,0 @@
-package tracetestcli_test
-
-import (
-	"testing"
-
-	"github.com/kubeshop/tracetest/cli-e2etest/tracetestcli"
-	"github.com/stretchr/testify/require"
-)
-
-func TestHelpCommand(t *testing.T) {
-	expectedExitCode := 0
-
-	possibleCommands := []string{"help", "--help", "-h"}
-
-	// test multiple help commands variations
-	for _, helpCommand := range possibleCommands {
-		t.Run(helpCommand, func(t *testing.T) {
-			result, err := tracetestcli.Exec(helpCommand)
-			require.NoError(t, err)
-
-			require.Equal(t, expectedExitCode, result.ExitCode)
-		})
-	}
-}
