+++ testing/cli-e2etest/tracetestcli/usecases/version_test.go
@@ -1,4 +1,4 @@
-package tracetestcli_test
+package usecases
 
 import (
 	"testing"
@@ -8,7 +8,12 @@ import (
 )
 
 func TestVersionCommand(t *testing.T) {
+	// Given I am a Tracetest CLI user
+	// When I try to check the tracetest version
+	// Then I should receive a version string with sucess
+
 	result, err := tracetestcli.Exec("version")
 	require.NoError(t, err)
 	require.Equal(t, 0, result.ExitCode)
+	require.Greater(t, len(result.StdOut), 0)
 }
